blog/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
black==20.8b1
click==7.1.2
Django==3.1.7
django-taggit==1.3.0
Faker==6.6.0
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2-binary==2.8.6
pylint==2.7.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
markdown==3.2.1
